requirements.txt
aws-cdk-lib==2.63.2
constructs>=10.0.0,<11.0.0
boto3
PyYAML
